+++ SabiMarket.API/ServiceExtensions/ServiceCollectionExtensions.cs
@@ -161,7 +161,8 @@ public static class ServiceCollectionExtensions
         services.AddScoped<IAuthenticationService, AuthenticationService>();
         services.AddScoped<IValidator<RegistrationRequestDto>, RegistrationRequestValidator>();
         services.AddScoped<IValidator<LoginRequestDto>, LoginRequestValidator>();
-        services.AddTransient<IEmailService, FreeEmailService>();
+        // services.AddTransient<IEmailService, FreeEmailService>();
+        services.AddSingleton<IEmailService, MailjetEmailService>();
         services.AddScoped<IRepositoryManager, RepositoryManager>();
         services.AddScoped<IServiceManager, ServiceManager>();
         services.AddScoped<ICloudinaryService, CloudinaryService>();
