+++ SabiMarket.Infrastructure/SabiMarket.Infrastructure.csproj
@@ -17,6 +17,7 @@
     <PackageReference Include="AutoMapper" Version="13.0.1" />
     <PackageReference Include="EPPlus" Version="7.5.3" />
     <PackageReference Include="itext" Version="8.0.1" />
+    <PackageReference Include="Mailjet.Api" Version="3.0.0" />
     <PackageReference Include="MailKit" Version="4.10.0" />
     <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="8.0.11" />
